# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Ronda API'
  contact:
    email: 'hello@andreacanton.dev'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
servers:
  - url: https://api.rondadellacaritaverona.org
    description: Production
  - url: http://localhost:3000
    description: Development
paths:
...
  /recipients:
    get:
      tags:
        - 'recipient'
      summary: 'Get list of recipients'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: 'query'
          name: 'q'
          type: 'string'
          description: 'Search parameter on email, memberNumber, firstname and lastname'
        - in: 'query'
          name: 'p'
          schema:
            type: 'integer'
          description: 'Page number. Default 1'
        - in: 'query'
          name: 'psize'
          schema:
            type: 'integer'
          description: 'Page size. Default 20'
        - in: 'query'
          name: 'status'
          schema:
            type: 'string'
            enum:
              - 'enabled'
              - 'disabled'
          description: 'Filter per Status'
        - in: 'query'
          name: 'sort'
          description: 'Field name to sort with'
          schema:
            type: 'string'
        - in: 'query'
          name: 'dir'
          description: 'Sort direction. Default `asc`'
          schema:
            type: 'string'
            enum:
              - 'asc'
              - 'desc'
      responses:
        '200':
          description: 'successful operation'
          schema:
            schema:
            type: object
            properties:
              models:
                type: array
                items:
                  $ref: '#/components/schemas/Recipient'
              pagination:
                type: object
                properties:
                  page:
                    type: number
                  pageSize:
                    type: number
                  rowCount:
                    type: number
                  pageCount:
                    type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - Bearer: []
    post:
      tags:
        - 'recipient'
      summary: 'Register recipient'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - name: 'recipient'
          in: 'body'
          schema:
            $ref: '#/components/requestBody/Recipient'
      security:
        - Bearer: []
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/Recipient'
            properties:
              message:
                type: 'string'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /recipients/{id}:
    parameters:
      - in: path
        name: id
        type: integer
    get:
      tags:
        - 'recipient'
      summary: 'Get recipient'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/Recipient'
            properties:
              message:
                type: 'string'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
        - 'recipient'
      summary: 'Edit recipient'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - name: 'recipient'
          in: 'body'
          schema:
            $ref: '#/components/requestBody/Recipient'
      security:
        - Bearer: []
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/Recipient'
            properties:
              message:
                type: 'string'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - 'recipient'
      summary: 'Deactivate recipient'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: object
            properties:
              message:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /recipients/is-card-number-taken/{fieldValue}:
    get:
      tags:
        - 'recipient'
      parameters:
        - name: fieldValue
          in: path
          schema:
            type: string
        - name: recipientId
          in: query
          schema:
            type: string
      produces:
        - 'application/json'
      security:
        - Bearer: []
      responses:
        '200':
          schema:
            type: object
            properties:
              isTaken:
                type: boolean
...
  /orders:
    get:
      tags:
        - 'orders'
      summary: 'Get list of orders'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      parameters:
        - in: 'query'
          name: 'q'
          type: 'string'
          description: 'Search parameter on email, memberNumber, firstname and lastname'
        - in: 'query'
          name: 'p'
          schema:
            type: 'integer'
          description: 'Page number. Default 1'
        - in: 'query'
          name: 'psize'
          schema:
            type: 'integer'
          description: 'Page size. Default 20'
        - in: 'query'
          name: 'status'
          schema:
            type: 'string'
            enum:
              - 'enabled'
              - 'disabled'
          description: 'Filter per Status'
        - in: 'query'
          name: 'sort'
          description: 'Field name to sort with'
          schema:
            type: 'string'
        - in: 'query'
          name: 'dir'
          description: 'Sort direction. Default `asc`'
          schema:
            type: 'string'
            enum:
              - 'asc'
              - 'desc'
      responses:
        '200':
          description: 'successful operation'
          schema:
            schema:
            type: object
            properties:
              models:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
              pagination:
                type: object
                properties:
                  page:
                    type: number
                  pageSize:
                    type: number
                  rowCount:
                    type: number
                  pageCount:
                    type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - 'orders'
      summary: 'Create order'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      parameters:
        - name: 'order'
          in: 'body'
          schema:
            $ref: '#/components/requestBody/Order'
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{id}:
    get:
      tags:
        - 'orders'
      summary: 'Get order'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
        - 'orders'
      summary: 'Edit order'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      parameters:
        - name: 'order'
          in: 'body'
          schema:
            $ref: '#/components/requestBody/Order'
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{id}/prepare:
    patch:
      tags:
        - 'orders'
      summary: 'Prepare order'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      parameters:
        - name: 'order'
          in: 'body'
          schema:
            type: object
            properties:
              notes:
                type: string
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /orders/{id}/cancel:
    patch:
      tags:
        - 'orders'
      summary: 'Cancel order'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      security:
        - Bearer: []
      parameters:
        - name: 'order'
          in: 'body'
          schema:
            type: object
            properties:
              notes:
                type: string
      responses:
        '200':
          description: 'successful operation'
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  requestBody:
...
    Recipient:
      type: object
      properties:
        cardNumber:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        gender:
          type: string
        topSize:
          type: string
        bottomSize:
          type: string
        shoeSize:
          type: string
    Order:
      type: object
      properties:
        recipientId:
          type: integer
        destination:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: integer
              size:
                type: string
              genre:
                type: string
              quantity:
                type: integer
              notes:
                type: string
  schemas:
...
    Recipient:
      type: 'object'
      properties:
        recipientId:
          type: integer
        cardNumber:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        gender:
          type: string
        topSize:
          type: string
        bottomSize:
          type: string
        shoeSize:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    Item:
      type: object
      properties:
        itemId:
          type: integer
        name:
          type: string
        sizeCategory:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    Order:
      type: object
      properties:
        orderId:
          type: integer
        orderNumber:
          type: string
        recipientId:
          type: integer
        destination:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        recipient:
          $ref: '#/components/schemas/Recipient'
    OrderDetail:
      type: object
      properties:
        orderId:
          type: integer
        orderNumber:
          type: string
        recipientId:
          type: integer
        destination:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        recipient:
          type: object
          properties:
            cardNumber:
              type: string
            firstname:
              type: string
            lastname:
              type: string
        orderNotes:
          type: array
          items:
            type: object
            properties:
              orderNoteId:
                type: integer
              orderId:
                type: integer
              userId:
                type: string
              phase:
                type: string
              body:
                type: string
              createdAt:
                type: string
              updatedAt:
                type: string
              user:
                type: object
                properties:
                  firstname:
                    type: string
                  lastname:
                    type: string
        orderItems:
          type: array
          items:
            type: object
            properties:
              orderItemId:
                type: integer
              orderId:
                type: integer
              itemId:
                type: integer
              size:
                type: string
              genre:
                type: string
              quantity:
                type: integer
              item:
                $ref: '#/components/schemas/Item'
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message